fix(Emacs): move org-mode sooner in the load chain
I kept running into issues with errors like "Symbol’s function
definition is void: org-file-name-concat" when loading and using
`org-mode` buffers. According to [an issue on GitHub], the fix is
moving the requiring/loading of `org-mode` to earlier in the process
before any other package has a chance to load it.

[an issue on GitHub]: https://github.com/org-roam/org-roam/issues/1916#issuecomment-1013682835